+++ .env
@@ -1,2 +1,2 @@
-NEXT_PUBLIC_BASE_URL=http://122.45.170.200:8080
-NEXT_PUBLIC_TOKEN=Bearer eyJyZWdEYXRlIjoxNzI4MjYzMTM2MDQ4LCJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJkMzgwZTI5ZC03YWMwLTQ0ZTUtOGJhZC1mMzY1OTVkNTVlZjYiLCJpZCI6NzgsImlhdCI6MTcyODI2MzEzNiwiZXhwIjoxNzM2MjExOTM2fQ.k21U6CzA9wsuJ-ctmhjaj0xqR54bNxgABjHhxo18eXY
+NEXT_PUBLIC_BASE_URL=https://api.suldak.co.kr/app
+NEXT_PUBLIC_TOKEN=Bearer eyJyZWdEYXRlIjoxNzM2NDAyNDA3NDI2LCJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiJzdWxkYWtAbWUuY29tIiwidXNlcklkIjoyLCJ1c2VyRW1haWwiOiJzdWxkYWtAbWUuY29tIiwibmlja25hbWUiOiJzdHJpbmciLCJpYXQiOjE3MzY0MDI0MDcsImV4cCI6MTc0NDE3ODQwN30.QIk8XNQ5OYtVTfvZaRiTiCGg5SBHC0ukcD6smJj0lMs
