+++ components/liquor/category/LiquorCategoryContent.tsx
@@ -1,0 +1,61 @@
+import { Liquor } from "models/liquor";
+import { useLiquorSearch } from "apis/liquor/useLiquorSearch";
+import CategoryHeader from "components/liquor/category/CategoryHeader";
+import NoResultSection from "components/liquor/search/section/NoResultSection";
+import LoadingCard from "components/shared/LiquorCard/LoadingCard";
+import CategoryFilter from "components/liquor/category/CategoryFilter";
+import LiquorList from "components/liquor/category/LiquorList";
+import SearchInfoSection from "components/liquor/category/SearchInfoSection";
+
+function LiquorCategoryContent({
+  searchParams,
+}: {
+  searchParams: URLSearchParams;
+}) {
+  const query = searchParams.get("q");
+  const tag = !query || query === "전체" ? undefined : query;
+
+  const { data, isLoading, error } = useLiquorSearch(
+    {
+      tag,
+      recordSize: 100,
+    },
+    searchParams.toString(),
+  );
+
+  const liquors: Liquor[] = data?.data.content || [];
+  const tagValue = searchParams.get("q") || "";
+
+  if (error) {
+    return (
+      <main className="flex min-h-screen flex-col">
+        <CategoryHeader tagValue={"404"} />
+        <div className="flex flex-grow items-center justify-center">
+          <p>오류가 발생했습니다. 다시 시도해주세요.</p>
+        </div>
+      </main>
+    );
+  }
+
+  return (
+    <main className="flex min-h-screen flex-col pb-[10px]">
+      <CategoryHeader tagValue={tagValue} />
+      <CategoryFilter />
+      <SearchInfoSection count={isLoading ? 0 : liquors.length} />
+      {isLoading ? (
+        <section className="flex flex-col items-center justify-center gap-2.5 overflow-y-auto px-[20px]">
+          <LoadingCard />
+          <LoadingCard />
+          <LoadingCard />
+          <LoadingCard />
+        </section>
+      ) : liquors.length === 0 ? (
+        <NoResultSection />
+      ) : (
+        <LiquorList liquors={liquors} />
+      )}
+    </main>
+  );
+}
+
+export default LiquorCategoryContent;
